+++ R/estimamodelo.r
@@ -13,6 +13,8 @@
 #' \describe{
 #'     \item{\code{reg_lin}}{Regressao linear comum}
 #'     \item{\code{reg_quant}}{Regressao linear quantílica}
+#'     \item{\code{GAM}}{Modelos Aditivos Generalizados}
+#'     \item{\code{BOOST}}{Modelagem por boosting de modelos simples}
 #'     \item{\code{sarima}}{SARIMA(p, d, q)(P, D, Q)}
 #'     \item{\code{sarimax}}{SARIMAX(p, d, q)(P, D, Q)}
 #'     \item{\code{ss_ar1_saz}}{Espaço de estados composto por processo AR(1) + Sazonalidade}
@@ -136,10 +138,12 @@ estimamodelo.numeric <- function(serie, tipo, periodico = FALSE, ...) estimamode
 
 estimamodelo.ts <- function(serie, tipo, periodico = FALSE, ...) {
 
-    fitfunc <- ifelse(periodico, estimamodelo_P, estimamodelo_U)
-    out <- fitfunc(serie, tipo, ...)
+    fitfun <- ifelse(periodico, estimamodelo_P, estimamodelo_U)
+    mc <- match.call()
+    mc[[1]] <- fitfun
+    mc$periodico <- NULL
 
-    return(out)
+    eval(mc, parent.frame())
 }
 
 # METODOS -----------------------------------------------------------------------------------------