+++ R/janelamovel.r
@@ -82,17 +82,17 @@ janelamovel <- function(serie, tipo, janela, passo = 1L, n.ahead = 1L, refit.cad
 
     args <- list(...)
 
-    if(!is.ts(serie)) serie <- ts(serie)
+    if (!is.ts(serie)) serie <- ts(serie)
 
     has_regdata <- "regdata" %in% ...names()
-    if(!has_regdata) {
+    if (!has_regdata) {
         regdata <- NULL
     } else {
         regdata <- args$regdata
         args$regdata <- NULL
     }
 
-    if(has_regdata && (length(serie) + n.ahead > nrow(regdata))) {
+    if (has_regdata && (length(serie) + n.ahead > nrow(regdata))) {
         warning("'regdata' deve conter 'length(serie) + n.ahead' observacoes -- reduzindo 'serie'")
 
         reduz <- length(serie) + n.ahead - nrow(regdata)
@@ -115,7 +115,8 @@ janelamovel <- function(serie, tipo, janela, passo = 1L, n.ahead = 1L, refit.cad
 
     retfun <- whichreturn(full.output)
 
-    jm <- lapply(seq(janelas), function(i) {
+    jm <- vector("list", length(janelas))
+    for (i in seq_along(janelas)) {
 
         verb_func(janelas[[i]][[1]], janelas[[i]][[2]], v_refit[i])
 
@@ -128,8 +129,8 @@ janelamovel <- function(serie, tipo, janela, passo = 1L, n.ahead = 1L, refit.cad
         inewdata <- regdata[window(aux, ijn[[1]], ijn[[2]]), , drop = FALSE]
         pred <- predict(mod, n.ahead, newdata = inewdata)
 
-        retfun(pred, mod, inewdata)
-    })
+        jm[[i]] <- retfun(pred, mod, inewdata)
+    }
 
     # Retorna
     return(jm)
@@ -163,11 +164,13 @@ verbose_fun <- function(verbose) {
 
     verb_func <- switch(as.character(verbose),
         "0" = function(...) NULL,
-        "1" = function(i, f, r) if(r) cat("\t Prevendo serie [", i, "] -> [", f, "]\n") else NULL,
-        "2" = function(i, f, r) if(r) {
-            cat("REFIT -- Prevendo serie [", i, "] -> [", f, "]\n")
-        } else {
-            cat("\t Prevendo serie [", i, "] -> [", f, "]\n")
+        "1" = function(i, f, r) if (r) cat("\t Prevendo serie [", i, "] -> [", f, "]\n") else NULL,
+        "2" = function(i, f, r) {
+            if (r) {
+                cat("REFIT -- Prevendo serie [", i, "] -> [", f, "]\n")
+            } else {
+                cat("\t Prevendo serie [", i, "] -> [", f, "]\n")
+            }
         }
     )
 
@@ -198,8 +201,6 @@ verbose_fun <- function(verbose) {
 #' @return lista na qual cada elemento e uma lista de dois elementos, contendo o instante inicial e
 #'     final de \code{serie} a considerar em cada janela, no sistema de tempo \code{ts}. Deve ser 
 #'     notado que os instantes definem uma janela fechada no inicio e aberta no final
-#' 
-#' @importFrom utils tail
 
 expandejanelas <- function(serie, janela, passo) {
 
@@ -207,7 +208,7 @@ expandejanelas <- function(serie, janela, passo) {
     S <- frequency(serie)
     INI <- start(serie)
 
-    if(length(janela) > 1) {
+    if (length(janela) > 1) {
         inifix <- TRUE
     } else {
         inifix <- FALSE
@@ -217,9 +218,9 @@ expandejanelas <- function(serie, janela, passo) {
     v_ends <- seq(janela[2], N - (janela[1] - 1), by = passo)
 
     ultimo <- N - (janela[1] - 1)
-    if(tail(v_ends, 1) != ultimo) v_ends <- c(v_ends, ultimo)
+    if (tail(v_ends, 1) != ultimo) v_ends <- c(v_ends, ultimo)
 
-    if(inifix) {
+    if (inifix) {
         ini <- deltats(INI, janela[1] - 1, S)
         janelas <- lapply(v_ends, function(i) {
             end <- deltats(ini, i - 1, S)
@@ -253,9 +254,9 @@ expandejanelas <- function(serie, janela, passo) {
 expanderefit <- function(janelas, refit.cada) {
 
     nj <- length(janelas)
-    if((length(refit.cada) == 1) && is.na(refit.cada)) refit.cada <- length(janelas)
+    if ((length(refit.cada) == 1) && is.na(refit.cada)) refit.cada <- length(janelas)
 
-    if(length(refit.cada) == 1) {
+    if (length(refit.cada) == 1) {
         v_refit <- seq(1, nj, by = refit.cada)
     } else {
         stop("'refit.cada' como vetor de indices ainda nao implementado")