+++ R/mod-ss_ar1_saz.r
@@ -25,18 +25,18 @@ NULL
 
 ss_ar1_saz <- function(serie, ...) {
 
-    Z <- matrix(c(1, 1), 1)
-    T <- matrix(c(1, 0, 0, 0), 2)
-    R <- matrix(c(0, 1), 2)
+    arrZ <- matrix(c(1, 1), 1)
+    arrT <- matrix(c(1, 0, 0, 0), 2)
+    arrR <- matrix(c(0, 1), 2)
 
-    if(frequency(serie) == 1) {
+    if (frequency(serie) == 1) {
         mod <- SSModel(serie ~ -1 +
-            SSMcustom(Z = Z, T = T, R = R, a1 = c(1, 0), Q = NA),
+                SSMcustom(Z = arrZ, T = arrT, R = arrR, a1 = c(1, 0), Q = NA),
             H = 0)
     } else {
         mod <- SSModel(serie ~ -1 +
-            SSMcustom(Z = Z, T = T, R = R, a1 = c(1, 0), Q = NA) +
-            SSMseasonal(period = frequency(serie), sea.type = "dummy", Q = NA),
+                SSMcustom(Z = arrZ, T = arrT, R = arrR, a1 = c(1, 0), Q = NA) +
+                SSMseasonal(period = frequency(serie), sea.type = "dummy", Q = NA),
             H = 0)
     }
     upfunc <- function(par, model) {
@@ -49,7 +49,7 @@ ss_ar1_saz <- function(serie, ...) {
     }
     fit <- fitSSM(mod, inits = c(mean(serie), 0, 0, 0), updatefn = upfunc, method = "BFGS")
 
-    if(abs(logLik(fit$model)) < 1e-10) {
+    if (abs(logLik(fit$model)) < 1e-10) {
         fit$model$Z[] <- NA
     }
 
@@ -110,20 +110,20 @@ predict.ss_ar1_saz <- function(object, n.ahead, ...) {
 
 update.ss_ar1_saz <- function(object, newseries, refit = FALSE, ...) {
 
-    if(refit) {
+    if (refit) {
         object <- estimamodelo(newseries, "ss_ar1_saz")
     } else {
         modelo <- object$modelo
 
         # Se modelo não convergiu, tenta reestimar
-        if(all(is.na(modelo$Z))) return(estimamodelo(newseries, tipo = "ss_ar1_saz")$modelo)
+        if (all(is.na(modelo$Z))) return(estimamodelo(newseries, tipo = "ss_ar1_saz")$modelo)
 
         # Do contrario, atualiza normalmente
         modelo$y <- newseries
         attr(modelo$y, "dim") <- c(length(newseries), 1)
         attr(modelo, "n") <- as.integer(length(newseries))
 
-        newseries  <- if(is.ts(newseries)) newseries else ts(newseries)
+        newseries  <- if (is.ts(newseries)) newseries else ts(newseries)
 
         object <- new_modprevU(modelo, newseries, "ss_ar1_saz")
     }